% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPD_SIS_LRM.R
\name{dpd.sis.lrm}
\alias{dpd.sis.lrm}
\title{DPD-SIS under a Linear Regression Model with unknwon error variance}
\usage{
dpd.sis.lrm(
  d,
  y,
  X,
  alpha,
  Method = "L-BFGS-B",
  p = ncol(X),
  Initial = matrix(rep(1, p))
)
}
\arguments{
\item{d}{Integer. A list of the data matrices. Number of features to be selected. (1<= d <=p)}

\item{y}{Vector.   The response vector [n X 1].}

\item{X}{Matrix. Covariate Matrix  [n X p].}

\item{alpha}{Numeric. The DPD tuning parameter (0<= alpha <=1)}

\item{Method}{String. Numerical optimization method to be used for computation of marginal slopes. Possible options are "L-BFGS-B","Nelder-Mead", "BFGS", "CG", which are the same as the input of 'optim' function in R. Optional. Default is "L-BFGS-B".}

\item{p}{Integer. Number of columns in the covariate matrix (X). Optional.}

\item{Initial}{Vector. Initial values of the marginal slope parameter for estimation process. Default: 1 for all parameters. Optional.}
}
\value{
SIS Vector. A d X 2 vector where the first column contains the variable index (ranked in decreasing order of importance) and the second column consist of the corresponding MDPDE of the slope
}
\description{
Performs the Density Power Divergence based Robust Variable Screening (DPD-SIS) under a Linear Regression Model  y=X*beta + e, with e ~ N(0, sigma^2) for some unknown sigma, using parallel computation.
}
\details{
Reference: Ghosh, A. and Thoresen, M. (2021). A Robust Variable Screening procedure for Ultra-high dimensional data. Statistical Methods in Medical Research, 30(8), 1816–1832.
}
\examples{
\donttest{
n <- 50; p <- 500;
beta <- rep(0,p); beta[c(1:5)] <- c(1,1,1,1,1);
d <- floor(n/log(n));
Sigma <- diag(p-1);
X <- mvrnorm(n, mu=rep(0,p-1), Sigma=Sigma);
X0 <- cbind(1,X)
Y <- drop(X0 \%*\% beta + 2*rnorm(n))

alpha <- 0.3
SIS<-dpd.sis.lrm(d,Y,X0,alpha)
}

}
