% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdpd.r
\name{lmdpd}
\alias{lmdpd}
\title{MDPDE for Linear Regression Model}
\usage{
lmdpd(y, X, alpha, method = "L-BFGS-B", p = ncol(X), Initial = matrix(c(median(y), rep(1, p - 1), log(mad(y)))))
}
\arguments{
\item{y}{Vector.   The response vector [n X 1].}

\item{X}{Matrix. Covariate Matrix  [n X p].}

\item{alpha}{Numeric. The DPD tuning parameter (0<= alpha <=1)}

\item{p}{Integer. Number of columns in the covariate matrix (X). Optional.}

\item{Initial}{Vector. Initial values of the parameters for the estimation process (initial of sigma needs to be given via log(sigma)). Optional. Default is 1 for all slope parameters, median of y for intercept and MAD(y) for sigma.}

\item{Method}{String. Numerical optimization method to be used for the estimation process. Possible options are "L-BFGS-B","Nelder-Mead", "BFGS", "CG", which are the same as the input of 'optim' function in R. Optional. Default is "L-BFGS-B".}
}
\value{
A List of two elements.
The first list element is a [(p+1) X 1] vector containing the parameter estimates (last entry being estimate of sigma)
The second list element is an indicator if the numerical optimization within the estimation process has converged: 0 == convergence. (Same as the convergence from 'optim' function of R)
}
\description{
Computes the robust Minimum Density Power Divergence Estimator (MDPDE) under a Linear Regression Model  y=X*beta + e, with e ~ N(0, sigma^2) for unknown sigma.
}
\details{
Reference: Ghosh, A., & Basu, A. (2013). Robust estimation for independent non-homogeneous observations using density power divergence with applications to linear regression. Electronic Journal of statistics, 7, 2420-2456.
}
\examples{
\donttest{
n <- 50; p <- 5;
beta <- rep(1,p); sigma<-.5
SigmaX <- diag(p-1);
X <- mvrnorm(n, mu=rep(0,p-1), Sigma=SigmaX);
X0 <- cbind(1,X)
Y <- drop(X0 \%*\% beta + sigma*rnorm(n))

alpha <- 0.3
MDPDE<-lmdpd(Y,X0,alpha)
}

}
