% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Competitor_robSIS_LRM.R
\name{MCP.sis.lrm}
\alias{MCP.sis.lrm}
\title{MCP correlation based SIS under a Linear Regression Model}
\usage{
MCP.sis.lrm(d, y, X)
}
\arguments{
\item{d}{Integer. A list of the data matrices. Number of features to be selected. (1<= d <=p)}

\item{y}{Vector.   The response vector [n X 1].}

\item{X}{Matrix. Covariate Matrix  [n X p].}
}
\value{
SIS Vector. A d X 2 vector where the first column contains the variable index (ranked in decreasing order of importance) and the second column consist of the corresponding MDPDE of the slope
}
\description{
Performs the MCP correlation based Robust Variable Screening (DPD-SIS) under a Linear Regression Model  y=X*beta + e, with e ~ N(0, sigma^2) for some unknown sigma, using parallel computation.
}
\details{
Reference: Mu W, Xiong S.  Some notes on robust sure independence screening.  {J App Stat}. 2014; 41(10):2092--2102.
}
\examples{
\donttest{
n <- 50; p <- 500;
beta <- rep(0,p); beta[c(1:5)] <- c(1,1,1,1,1);
d <- floor(n/log(n));
Sigma <- diag(p-1);
X <- mvrnorm(n, mu=rep(0,p-1), Sigma=Sigma);
X0 <- cbind(1,X)
Y <- drop(X0 \%*\% beta + 2*rnorm(n))

SIS<-MCP.sis.lrm(d,Y,X0)
}

}
